git-svn-id: file:///home/cysun/subversion/gefp/web/trunk@4 1e33301c-9783-4aea-b6a3-3f25a6be623f

+++ src/main/java/gefp/model/dao/FlightPlanDao.java
@@ -5,10 +5,10 @@ import gefp.model.FlightPlan;
 import java.util.List;
 
 public interface FlightPlanDao {
-	
-	FlightPlan getFlightPlan( Integer id );
+
+    FlightPlan getFlightPlan( Long id );
 
     List<FlightPlan> getFlightPlans();
-    
-    FlightPlan saveFlightPlan(FlightPlan flightplan);
+
+    FlightPlan saveFlightPlan( FlightPlan flightplan );
 }
